+++ custom_components/sberhome/www/components/sberhome-debug-view.js
@@ -7,11 +7,7 @@
  *    sberhome.send_raw_command + показ response
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 const PRESETS = [
   {