+++ custom_components/sberhome/www/components/sberhome-device-modal.js
@@ -8,11 +8,7 @@
  *  - Raw JSON (для багрепортов)
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 const IMG_BASE = "https://img.iot.sberdevices.ru";