+++ custom_components/sberhome/www/components/sberhome-toast.js
@@ -3,11 +3,7 @@
  * Adapted from MQTT-SberGate sber-toast.js.
  */
 
-const LitElement = Object.getPrototypeOf(
-  customElements.get("ha-panel-lovelace") ?? customElements.get("hui-view")
-);
-const html = LitElement?.prototype.html;
-const css = LitElement?.prototype.css;
+import { LitElement, html, css } from "../lit-base.js";
 
 class SberHomeToast extends LitElement {
   static get properties() {